Bug almost solved. Reason for the periodically shifted functions was a special property of the DFT and the index shift of half the box. This effect was solved by the new implemented shift_res method in StaticCalcs.hpp.
Testing the analytical result with the time evolution operator
revealed, that imaginary and real part are exchanged. Thus, there's still
an error!

+++ include/math/FourierCoeff.hpp
@@ -31,7 +31,7 @@ namespace math {
         
         for(int i = 0; i < nx; i++)
         {   
-            x = x0 + i * h;
+            x = x0 + double(i) * h;
             res += psi(x)*exp(-2*M_PI*m*iu*double(i)/nx1);
             
         }
